re factorize common mixins
