+++ geotrek/authent/admin.py
@@ -8,7 +8,7 @@ from django.contrib.auth.models import User
 
 from geotrek.authent.models import Structure
 from geotrek.authent.models import UserProfile
-from geotrek.common.mixins import MergeActionMixin
+from geotrek.common.mixins.actions import MergeActionMixin
 
 
 admin.site.unregister(User)
