+++ geotrek/feedback/models.py
@@ -8,8 +8,7 @@ from django.db.models.query_utils import Q
 from django.utils.formats import date_format
 from django.utils.translation import gettext_lazy as _
 
-from geotrek.common.mixins import (AddPropertyMixin, NoDeleteMixin, PicturesMixin,
-                                   TimeStampedModelMixin)
+from geotrek.common.mixins.models import AddPropertyMixin, NoDeleteMixin, PicturesMixin, TimeStampedModelMixin
 from geotrek.common.utils import intersecting
 from geotrek.core.models import Path
 from geotrek.maintenance.models import Intervention
