+++ geotrek/maintenance/views.py
@@ -9,7 +9,7 @@ from mapentity.views import (MapEntityLayer, MapEntityList, MapEntityJsonList, M
                              MapEntityDetail, MapEntityDocument, MapEntityCreate, MapEntityUpdate, MapEntityDelete)
 
 from geotrek.altimetry.models import AltimetryMixin
-from geotrek.common.mixins import CustomColumnsMixin
+from geotrek.common.mixins.views import CustomColumnsMixin
 from geotrek.common.views import FormsetMixin
 from geotrek.authent.decorators import same_structure_required
 from .models import Intervention, Project, ManDay
