+++ geotrek/signage/models.py
@@ -8,7 +8,8 @@ from django.conf import settings
 from mapentity.models import MapEntityMixin
 
 from geotrek.authent.models import StructureOrNoneRelated
-from geotrek.common.mixins import AddPropertyMixin, OptionalPictogramMixin, NoDeleteManager
+from geotrek.common.mixins.models import AddPropertyMixin, OptionalPictogramMixin
+from geotrek.common.mixins.managers import NoDeleteManager
 from geotrek.common.models import Organism
 from geotrek.common.utils import classproperty, format_coordinates, collate_c, spatial_reference
