+++ requirements.txt
@@ -95,7 +95,7 @@ django-compressor==2.4
     # via mapentity
 django-crispy-forms==1.13.0
     # via mapentity
-django-debug-toolbar==3.2.1
+django-debug-toolbar==3.2.4
     # via geotrek (setup.py)
 django-embed-video==1.3.1
     # via
